+++ src/main/java/org/perscholas/capstone/postrocket/services/UserService.java
@@ -10,10 +10,10 @@ import java.util.List;
 
 public interface UserService extends UserDetailsService {
     List<User> getAllUsers();
-    User getUserById(int id);
+    User getUserById(long id);
 
     User getUserByEmail(String email);
-    void deleteUser(int id);
+    void deleteUser(long id);
 
     void updateUser(User user);
